backend/requirements.txt

Django>=4.2,<5.0
djangorestframework>=3.12.4
openai>=0.27.0
requests>=2.25.1
django-cors-headers
psycopg2-binary==2.9.9